RC Stage; Added totals; Added error handlers; Added password change (untested); Fixed score display for some; Fixed smtp auth in logcat sender; Some fixes and corrections;

+++ app/src/main/java/hexlay/ums/models/subject/SubjectDetail.kt
@@ -11,9 +11,9 @@ data class SubjectDetail(
     @SerializedName("type")
     var detailType: String,
     @SerializedName("minValue")
-    var detailMinScore: Int,
+    var detailMinScore: Double,
     @SerializedName("absolute")
-    var detailMaxScore: Int,
+    var detailMaxScore: Double,
     @SerializedName("grade")
     var detailGrade: SubjectGrade
 ) : Parcelable
